separate jax imports and function from main modules
