+++ dolphin/Util/__init__.py
@@ -1,0 +1,2 @@
+# -*- coding: utf-8 -*-
+"""Utility functions used for special cases in dolphin"""
